Fixed hash functions: when the key was lower than 0 (could happen with entry_refs and node_refs passed in) they returned an invalid index (larger than range). Fixed rootfs compilation with tracing turned on. Minor cleanup.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@14065 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/system/kernel/fs/bootfs.c
@@ -90,7 +90,7 @@ bootfs_vnode_hash_func(void *_v, const void *_key, uint32 range)
 	if (v != NULL)
 		return v->id % range;
 
-	return (*key) % range;
+	return (uint64)*key % range;
 }
